requirements.txt
streamlit
yt-dlp
pillow